pyproject.toml
[project]
name = "limit-order-algo-py"
version = "0.1.0"
description = "This project creates an algorithm to bid for the demand load volume of Tibber's customers EV fleet. It was prepared by Francisco Botero as part of the application to the vacancy as Quantitative Trader at Tibber."
authors = [
    {name = "Francisco Botero",email = "francisco.botero.bernal@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "ipykernel (>=6.30.1,<7.0.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "astral (>=3.2,<4.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "hyperopt (>=0.2.7,<0.3.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]

[tool.poetry]
packages = [{include = "limit_order_algo", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
